OBPIH-6097 Fixes after review
